# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Document APIs",
        "contact": {}
    },
    "paths": {
...
        "/cms/v1/coins/exchanges/{exchange_id}/trades": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "List Trades",
                "parameters": [
                    {
                        "enum": [
                            "binance",
                            "bitget",
                            "bybit",
                            "cex",
                            "gate",
                            "mexc",
                            "probit"
                        ],
                        "type": "string",
                        "description": "exchange_id",
                        "name": "exchange_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "sandbox",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/coin.Trade"
                            }
                        }
                    }
                }
            }
        },
...
        "/cms/v1/coins/traders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "List Traders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.TraderBaseDto"
                            }
                        }
                    }
                }
            }
        },
...
        "/cms/v1/traders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "List Traders",
                "parameters": [
                    {
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "DataStatusEnable",
                            "DataStatusDisable"
                        ],
                        "name": "data_status",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at.desc",
                        "name": "sorts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.TraderCmsDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Create Trader",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.TraderCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.TraderCmsDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    }
                }
            }
        },
        "/cms/v1/traders/{trader_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Update Trader",
                "parameters": [
                    {
                        "type": "string",
                        "description": "trader_id",
                        "name": "trader_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.TraderCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.TraderCmsDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "coin.Fee": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "rate": {
                    "type": "number"
                }
            }
        },
...
        "coin.Trade": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "cost": {
                    "type": "number"
                },
                "datetime": {
                    "type": "string"
                },
                "fee": {
                    "$ref": "#/definitions/coin.Fee"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "side": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "taker_or_maker": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "db.TraderBaseDto": {
            "type": "object",
            "properties": {
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "exchange": {
                    "$ref": "#/definitions/enum.Exchange"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "sandbox": {
                    "type": "boolean"
                },
                "trader_id": {
                    "type": "string",
                    "example": "671db9eca1f1b1bdbf3d4611"
                }
            }
        },
        "db.TraderCmsData": {
            "type": "object",
            "required": [
                "api_key",
                "data_status",
                "exchange",
                "name",
                "secret"
            ],
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "exchange": {
                    "$ref": "#/definitions/enum.Exchange"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "password": {
                    "type": "string"
                },
                "sandbox": {
                    "type": "boolean"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "db.TraderCmsDto": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string",
                    "example": "api_key"
                },
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "exchange": {
                    "$ref": "#/definitions/enum.Exchange"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "sandbox": {
                    "type": "boolean"
                },
                "secret": {
                    "type": "string",
                    "example": "secret"
                },
                "trader_id": {
                    "type": "string",
                    "example": "671db9eca1f1b1bdbf3d4611"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "editor"
                }
            }
        },
...
        "ecode.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "error_description": {
                    "type": "string"
                }
            }
        },
...
        "enum.DataStatus": {
            "type": "string",
            "enum": [
                "enable",
                "disable"
            ],
            "x-enum-varnames": [
                "DataStatusEnable",
                "DataStatusDisable"
            ]
        },
        "enum.Exchange": {
            "type": "string",
            "enum": [
                "binance",
                "bitget",
                "bybit",
                "cex",
                "gate",
                "mexc",
                "probit"
            ],
            "x-enum-varnames": [
                "ExchangeBinance",
                "ExchangeBitget",
                "ExchangeBybit",
                "ExchangeCex",
                "ExchangeGate",
                "ExchangeMexc",
                "ExchangeProbit"
            ]
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Header"
        }
    }
}